refactor: organize files better into proper folders as needed; update docs
