+++ src/transports.ts
@@ -8,8 +8,8 @@
 // Re-export all transport classes
 export { ConsoleTransport } from './transports/console';
 export { FileTransport, WorkerFileTransport } from './transports/file';
-export { AsyncFileTransport } from './transports/AsyncFileTransport';
-export { SyncFileTransport } from './transports/SyncFileTransport';
+export { AsyncFileTransport } from './transports/base/implementations/AsyncFileTransport';
+export { SyncFileTransport } from './transports/base/implementations/SyncFileTransport';
 export { StreamTransport } from './transports/stream';
 export { HTTPTransport } from './transports/http';
 export { NullTransport } from './transports/null';
