+++ src/types/external/aws-sdk.d.ts
