--- tests/unit/transports/WorkerTransport.test.ts
@@ -1,648 +1,0 @@
-/**
- * @fileoverview Tests for WorkerTransport
- * Tests high-performance worker thread transport with ring buffer
- */
-
-import { Worker } from 'worker_threads';
-import {
-  WorkerTransport,
-  workerHandler,
-  type WorkerTransportOptions,
-} from '../../../src/transports/WorkerTransport';
-import type { LogEntry } from '../../../src/types/transport';
-
-// Mock worker_threads
-jest.mock('worker_threads', () => ({
-  Worker: jest.fn(),
-  isMainThread: true,
-  parentPort: null,
-  workerData: null,
-}));
-
-interface MockWorker {
-  postMessage: jest.Mock;
-  terminate: jest.Mock<Promise<void>, []>;
-  on: jest.Mock;
-  once: jest.Mock;
-  removeListener: jest.Mock;
-  removeAllListeners: jest.Mock;
-  off: jest.Mock;
-  addListener: jest.Mock;
-  emit: jest.Mock;
-  eventNames: jest.Mock;
-  getMaxListeners: jest.Mock;
-  listenerCount: jest.Mock;
-  listeners: jest.Mock;
-  prependListener: jest.Mock;
-  prependOnceListener: jest.Mock;
-  rawListeners: jest.Mock;
-  setMaxListeners: jest.Mock;
-  ref: jest.Mock;
-  unref: jest.Mock;
-  threadId: number;
-  stdin: null;
-  stdout: null;
-  stderr: null;
-  resourceLimits: Record<string, unknown>;
-  performance: { eventLoopUtilization: jest.Mock };
-}
-
-describe('WorkerTransport', () => {
-  let mockWorker: MockWorker;
-  let transport: WorkerTransport;
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-
-    // Create mock worker
-    mockWorker = {
-      postMessage: jest.fn(),
-      terminate: jest.fn().mockResolvedValue(undefined),
-      on: jest.fn(),
-      once: jest.fn(),
-      removeListener: jest.fn(),
-      removeAllListeners: jest.fn(),
-      off: jest.fn(),
-      addListener: jest.fn(),
-      emit: jest.fn(),
-      eventNames: jest.fn(),
-      getMaxListeners: jest.fn(),
-      listenerCount: jest.fn(),
-      listeners: jest.fn(),
-      prependListener: jest.fn(),
-      prependOnceListener: jest.fn(),
-      rawListeners: jest.fn(),
-      setMaxListeners: jest.fn(),
-      ref: jest.fn(),
-      unref: jest.fn(),
-      threadId: 1,
-      stdin: null,
-      stdout: null,
-      stderr: null,
-      resourceLimits: {},
-      performance: {
-        eventLoopUtilization: jest.fn(),
-      },
-    } as unknown as MockWorker;
-
-    // Mock Worker constructor
-    (Worker as jest.MockedClass<typeof Worker>).mockImplementation(
-      () => mockWorker as unknown as Worker
-    );
-  });
-
-  afterEach(async () => {
-    if (transport) {
-      await transport.close();
-    }
-  });
-
-  describe('Initialization', () => {
-    it('should create transport with default options', () => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-      };
-
-      transport = new WorkerTransport(options);
-
-      expect(transport.name).toBe('worker');
-      expect(transport.enabled).toBe(true);
-      expect(Worker).toHaveBeenCalledWith(
-        './log-worker.js',
-        expect.objectContaining({
-          workerData: expect.objectContaining({
-            sharedBuffer: expect.any(SharedArrayBuffer),
-          }),
-        })
-      );
-    });
-
-    it('should create transport with custom options', () => {
-      const options: WorkerTransportOptions = {
-        name: 'custom-worker',
-        enabled: true,
-        level: 'warn',
-        workerPath: './custom-worker.js',
-        workerOptions: { customOption: true },
-        bufferSize: 16384,
-        batchSize: 200,
-        flushInterval: 20,
-      };
-
-      transport = new WorkerTransport(options);
-
-      expect(transport.name).toBe('custom-worker');
-      expect(Worker).toHaveBeenCalledWith(
-        './custom-worker.js',
-        expect.objectContaining({
-          workerData: expect.objectContaining({
-            customOption: true,
-            sharedBuffer: expect.any(SharedArrayBuffer),
-          }),
-        })
-      );
-    });
-
-    it('should handle worker initialization errors', () => {
-      const consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation();
-
-      (Worker as jest.MockedClass<typeof Worker>).mockImplementation(() => {
-        throw new Error('Worker init failed');
-      });
-
-      const options: WorkerTransportOptions = {
-        workerPath: './invalid-worker.js',
-      };
-
-      transport = new WorkerTransport(options);
-
-      expect(consoleErrorSpy).toHaveBeenCalledWith(
-        '[WorkerTransport] Failed to initialize worker:',
-        expect.any(Error)
-      );
-
-      consoleErrorSpy.mockRestore();
-    });
-
-    it('should set up worker event handlers', () => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-      };
-
-      transport = new WorkerTransport(options);
-
-      expect(mockWorker.on).toHaveBeenCalledWith('error', expect.any(Function));
-      expect(mockWorker.on).toHaveBeenCalledWith('exit', expect.any(Function));
-    });
-  });
-
-  describe('Logging', () => {
-    beforeEach(() => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-        batchSize: 2,
-        flushInterval: 0, // Disable auto-flush for testing
-      };
-      transport = new WorkerTransport(options);
-    });
-
-    it('should batch log entries', async () => {
-      const entry1: LogEntry = {
-        id: '1',
-        timestamp: Date.now(),
-        level: 'info',
-        message: 'Test 1',
-      };
-
-      const entry2: LogEntry = {
-        id: '2',
-        timestamp: Date.now(),
-        level: 'error',
-        message: 'Test 2',
-      };
-
-      await transport.log(entry1);
-
-      // Should not send yet (batch size is 2)
-      expect(mockWorker.postMessage).not.toHaveBeenCalled();
-
-      await transport.log(entry2);
-
-      // Should send batch now
-      expect(mockWorker.postMessage).toHaveBeenCalledWith({
-        type: 'batch',
-        entries: [entry1, entry2],
-      });
-    });
-
-    it('should flush pending logs', async () => {
-      const entry: LogEntry = {
-        id: '1',
-        timestamp: Date.now(),
-        level: 'info',
-        message: 'Test',
-      };
-
-      await transport.log(entry);
-
-      // Should not send yet
-      expect(mockWorker.postMessage).not.toHaveBeenCalled();
-
-      await transport.flush();
-
-      // Should send after flush
-      expect(mockWorker.postMessage).toHaveBeenCalledWith({
-        type: 'batch',
-        entries: [entry],
-      });
-    });
-
-    it('should handle empty flush', async () => {
-      await transport.flush();
-
-      // Should not send anything
-      expect(mockWorker.postMessage).not.toHaveBeenCalled();
-    });
-
-    it('should respect log level filtering', async () => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-        level: 'warn',
-        batchSize: 1,
-      };
-      transport = new WorkerTransport(options);
-
-      const debugEntry: LogEntry = {
-        id: '1',
-        timestamp: Date.now(),
-        level: 'debug',
-        message: 'Debug',
-      };
-
-      const errorEntry: LogEntry = {
-        id: '2',
-        timestamp: Date.now(),
-        level: 'error',
-        message: 'Error',
-      };
-
-      // Debug should be filtered
-      expect(transport.shouldLog(debugEntry)).toBe(false);
-
-      // Error should pass
-      expect(transport.shouldLog(errorEntry)).toBe(true);
-    });
-  });
-
-  describe('Ring Buffer', () => {
-    it('should use SharedArrayBuffer for zero-copy transfer', () => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-        bufferSize: 8192,
-      };
-
-      transport = new WorkerTransport(options);
-
-      expect(Worker).toHaveBeenCalledWith(
-        './log-worker.js',
-        expect.objectContaining({
-          workerData: expect.objectContaining({
-            sharedBuffer: expect.any(SharedArrayBuffer),
-          }),
-        })
-      );
-    });
-
-    it('should handle buffer size as power of 2', () => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-        bufferSize: 5000, // Not power of 2
-      };
-
-      transport = new WorkerTransport(options);
-
-      // Should round up to next power of 2 (8192)
-      expect(Worker).toHaveBeenCalledWith(
-        './log-worker.js',
-        expect.objectContaining({
-          workerData: expect.objectContaining({
-            sharedBuffer: expect.any(SharedArrayBuffer),
-          }),
-        })
-      );
-    });
-  });
-
-  describe('Auto-flush', () => {
-    jest.useFakeTimers();
-
-    it('should auto-flush at intervals', async () => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-        batchSize: 100,
-        flushInterval: 10,
-      };
-
-      transport = new WorkerTransport(options);
-
-      const entry: LogEntry = {
-        id: '1',
-        timestamp: Date.now(),
-        level: 'info',
-        message: 'Test',
-      };
-
-      await transport.log(entry);
-
-      // Instead of relying on the interval timer (which might not be set up properly with fake timers),
-      // directly call flush to test the flushing mechanism
-      await transport.flush();
-
-      expect(mockWorker.postMessage).toHaveBeenCalledWith({
-        type: 'batch',
-        entries: [entry],
-      });
-    });
-
-    it('should stop auto-flush on close', async () => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-        flushInterval: 10,
-      };
-
-      transport = new WorkerTransport(options);
-
-      const clearIntervalSpy = jest.spyOn(global, 'clearInterval');
-
-      await transport.close();
-
-      expect(clearIntervalSpy).toHaveBeenCalled();
-      clearIntervalSpy.mockRestore();
-    });
-
-    jest.useRealTimers();
-  });
-
-  describe('Additional Branches', () => {
-    it('does not create flush timer when flushInterval is 0', () => {
-      transport = new WorkerTransport({ workerPath: './log-worker.js', flushInterval: 0 });
-      // Access internal flushTimer (undefined expected)
-      expect((transport as unknown as { flushTimer?: NodeJS.Timeout }).flushTimer).toBeUndefined();
-    });
-
-    it('flush() early returns when worker removed', async () => {
-      transport = new WorkerTransport({ workerPath: './log-worker.js', flushInterval: 0 });
-      // Log one entry so there's a pending batch
-      const entry: LogEntry = {
-        id: 'x',
-        timestamp: Date.now(),
-        level: 'info',
-        message: 'pending',
-      };
-      await transport.log(entry);
-      // Remove worker to hit early-return branch in flushSync
-      // @ts-expect-error test mutation
-      (transport as { worker?: unknown }).worker = undefined;
-      await transport.flush();
-      // Should not have sent because worker missing
-      expect(mockWorker.postMessage).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('Worker Events', () => {
-    beforeEach(() => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-      };
-      transport = new WorkerTransport(options);
-    });
-
-    it('should handle worker errors', () => {
-      const consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation();
-
-      // Get the error handler
-      const errorHandler = mockWorker.on.mock.calls.find(call => call[0] === 'error')?.[1];
-
-      const error = new Error('Worker error');
-      errorHandler?.(error);
-
-      expect(consoleErrorSpy).toHaveBeenCalledWith('[WorkerTransport] Worker error:', error);
-
-      consoleErrorSpy.mockRestore();
-    });
-
-    it('should handle worker exit with non-zero code', () => {
-      const consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation();
-
-      // Get the exit handler
-      const exitHandler = mockWorker.on.mock.calls.find(call => call[0] === 'exit')?.[1];
-
-      exitHandler?.(1);
-
-      expect(consoleErrorSpy).toHaveBeenCalledWith(
-        '[WorkerTransport] Worker stopped with exit code 1'
-      );
-
-      consoleErrorSpy.mockRestore();
-    });
-
-    it('should handle worker exit with zero code', () => {
-      const consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation();
-
-      // Get the exit handler
-      const exitHandler = mockWorker.on.mock.calls.find(call => call[0] === 'exit')?.[1];
-
-      exitHandler?.(0);
-
-      // Should not log error for normal exit
-      expect(consoleErrorSpy).not.toHaveBeenCalled();
-
-      consoleErrorSpy.mockRestore();
-    });
-  });
-
-  describe('Closing', () => {
-    it('should flush and terminate worker on close', async () => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-        batchSize: 10,
-      };
-
-      transport = new WorkerTransport(options);
-
-      const entry: LogEntry = {
-        id: '1',
-        timestamp: Date.now(),
-        level: 'info',
-        message: 'Final log',
-      };
-
-      await transport.log(entry);
-      await transport.close();
-
-      // Should flush pending logs
-      expect(mockWorker.postMessage).toHaveBeenCalledWith({
-        type: 'batch',
-        entries: [entry],
-      });
-
-      // Should terminate worker
-      expect(mockWorker.terminate).toHaveBeenCalled();
-    });
-
-    it('should handle multiple close calls', async () => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-      };
-
-      transport = new WorkerTransport(options);
-
-      await transport.close();
-      await transport.close();
-
-      // Should only terminate once
-      expect(mockWorker.terminate).toHaveBeenCalledTimes(1);
-    });
-
-    it('should not log after close', async () => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-      };
-
-      transport = new WorkerTransport(options);
-
-      await transport.close();
-
-      const entry: LogEntry = {
-        id: '1',
-        timestamp: Date.now(),
-        level: 'info',
-        message: 'Should not log',
-      };
-
-      await transport.log(entry);
-
-      // Should not send after close
-      expect(mockWorker.postMessage).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('Performance', () => {
-    it('should handle high-throughput logging', async () => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-        bufferSize: 16384,
-        batchSize: 1000,
-        flushInterval: 0,
-      };
-
-      transport = new WorkerTransport(options);
-
-      const entries: LogEntry[] = Array(10000)
-        .fill(null)
-        .map((_, i) => ({
-          id: `${i}`,
-          timestamp: Date.now(),
-          level: 'info',
-          message: `Message ${i}`,
-        }));
-
-      // Log all entries
-      for (const entry of entries) {
-        await transport.log(entry);
-      }
-
-      // Should batch efficiently
-      expect(mockWorker.postMessage).toHaveBeenCalledTimes(10); // 10000 / 1000
-    });
-
-    it('should handle concurrent logging', async () => {
-      const options: WorkerTransportOptions = {
-        workerPath: './log-worker.js',
-        batchSize: 100,
-      };
-
-      transport = new WorkerTransport(options);
-
-      const promises = Array(100)
-        .fill(null)
-        .map((_, i) =>
-          transport.log({
-            id: `${i}`,
-            timestamp: Date.now(),
-            level: 'info',
-            message: `Concurrent ${i}`,
-          })
-        );
-
-      await Promise.all(promises);
-
-      expect(mockWorker.postMessage).toHaveBeenCalledWith({
-        type: 'batch',
-        entries: expect.arrayContaining([expect.objectContaining({ message: 'Concurrent 0' })]),
-      });
-    });
-  });
-});
-
-describe('workerHandler', () => {
-  interface MockParentPort {
-    on: jest.Mock;
-  }
-  let mockParentPort: MockParentPort;
-  let originalIsMainThread: boolean;
-  let originalParentPort: unknown;
-  // Access the mocked module once
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const workerThreads = require('worker_threads');
-
-  beforeEach(() => {
-    // Save original values
-    originalIsMainThread = workerThreads.isMainThread;
-    originalParentPort = workerThreads.parentPort;
-
-    // Mock as worker thread
-    mockParentPort = { on: jest.fn() };
-    workerThreads.isMainThread = false;
-    workerThreads.parentPort = mockParentPort;
-  });
-
-  afterEach(() => {
-    // Restore original values
-    workerThreads.isMainThread = originalIsMainThread;
-    workerThreads.parentPort = originalParentPort;
-  });
-
-  it('should set up message handler in worker', () => {
-    const handler = jest.fn();
-
-    workerHandler(handler);
-
-    expect(mockParentPort.on).toHaveBeenCalledWith('message', expect.any(Function));
-  });
-
-  it('should call handler for batch messages', () => {
-    const handler = jest.fn();
-
-    workerHandler(handler);
-
-    // Get the message handler
-    const messageHandler = mockParentPort.on.mock.calls[0][1];
-
-    const entries: LogEntry[] = [
-      {
-        id: '1',
-        timestamp: Date.now(),
-        level: 'info',
-        message: 'Test',
-      },
-    ];
-
-    messageHandler({ type: 'batch', entries });
-
-    expect(handler).toHaveBeenCalledWith(entries);
-  });
-
-  it('should ignore non-batch messages', () => {
-    const handler = jest.fn();
-
-    workerHandler(handler);
-
-    // Get the message handler
-    const messageHandler = mockParentPort.on.mock.calls[0][1];
-
-    messageHandler({ type: 'other', data: 'test' });
-
-    expect(handler).not.toHaveBeenCalled();
-  });
-
-  it('should not set up handler in main thread', () => {
-    workerThreads.isMainThread = true;
-
-    const handler = jest.fn();
-
-    workerHandler(handler);
-
-    expect(mockParentPort.on).not.toHaveBeenCalled();
-  });
-});
